rename media_list_response to avoid confusion
